+++ src/MauiReactor.Canvas/Internals/DrawingContext.cs
@@ -1,0 +1,7 @@
+﻿using Microsoft.Maui.Graphics;
+
+namespace MauiReactor.Canvas.Internals
+{
+    public record DrawingContext(ICanvas Canvas, RectF DirtyRect);
+
+}
