--- src/MauiReactor.Canvas/Platforms/Tizen/PlatformClass1.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace MauiReactor.Canvas
-{
-    // All the code in this file is only included on Tizen.
-    //public class PlatformClass1
-    //{
-    //}
-}
