--- src/MauiReactor.Canvas/Platforms/Windows/PlatformClass1.cs
@@ -1,7 +1,0 @@
-﻿namespace MauiReactor.Canvas
-{
-    // All the code in this file is only included on Windows.
-    //public class PlatformClass1
-    //{
-    //}
-}
