--- src/MauiReactor.Canvas/Platforms/iOS/PlatformClass1.cs
@@ -1,7 +1,0 @@
-﻿namespace MauiReactor.Canvas
-{
-    // All the code in this file is only included on iOS.
-    //public class PlatformClass1
-    //{
-    //}
-}
